+++ mayan/apps/document_comments/admin.py
@@ -6,7 +6,7 @@ from .models import Comment
 @admin.register(Comment)
 class CommentAdmin(admin.ModelAdmin):
     date_hierarchy = 'submit_date'
-    list_display = ('document', 'submit_date', 'user', 'comment')
+    list_display = ('document', 'submit_date', 'user', 'text')
     list_filter = ('user',)
-    readonly_fields = ('document', 'submit_date', 'user', 'comment')
-    search_fields = ('comment',)
+    readonly_fields = ('document', 'submit_date', 'user', 'text')
+    search_fields = ('text',)