+++ mayan/apps/document_comments/forms.py
@@ -8,7 +8,7 @@ from .models import Comment
 class DocumentCommentDetailForm(DetailForm):
 
     class Meta:
-        fields = ('comment',)
+        fields = ('text',)
         extra_fields = (
             {'field': 'submit_date', 'widget': forms.widgets.DateTimeInput},
             {'field': 'user'},