+++ mayan/apps/document_comments/tests/test_views.py
@@ -113,7 +113,7 @@ class DocumentCommentViewTestCase(
 
         response = self._request_test_comment_detail_view()
         self.assertContains(
-            response=response, text=self.test_document_comment.comment,
+            response=response, text=self.test_document_comment.text,
             status_code=200
         )
 
@@ -133,13 +133,13 @@ class DocumentCommentViewTestCase(
     def test_comment_edit_view_no_permission(self):
         self._create_test_comment()
 
-        comment_text = self.test_document_comment.comment
+        comment_text = self.test_document_comment.text
 
         response = self._request_test_comment_edit_view()
         self.assertEqual(response.status_code, 404)
 
         self.test_document_comment.refresh_from_db()
-        self.assertEqual(self.test_document_comment.comment, comment_text)
+        self.assertEqual(self.test_document_comment.text, comment_text)
 
     def test_comment_edit_view_with_access(self):
         self._create_test_comment()
@@ -148,13 +148,13 @@ class DocumentCommentViewTestCase(
             obj=self.test_document, permission=permission_document_comment_edit
         )
 
-        comment_text = self.test_document_comment.comment
+        comment_text = self.test_document_comment.text
 
         response = self._request_test_comment_edit_view()
         self.assertEqual(response.status_code, 302)
 
         self.test_document_comment.refresh_from_db()
-        self.assertNotEqual(self.test_document_comment.comment, comment_text)
+        self.assertNotEqual(self.test_document_comment.text, comment_text)
 
     def test_trashed_document_comment_edit_view_with_access(self):
         self._create_test_comment()
@@ -163,7 +163,7 @@ class DocumentCommentViewTestCase(
             obj=self.test_document, permission=permission_document_comment_edit
         )
 
-        comment_text = self.test_document_comment.comment
+        comment_text = self.test_document_comment.text
 
         self.test_document.delete()
 
@@ -171,14 +171,14 @@ class DocumentCommentViewTestCase(
         self.assertEqual(response.status_code, 404)
 
         self.test_document_comment.refresh_from_db()
-        self.assertEqual(self.test_document_comment.comment, comment_text)
+        self.assertEqual(self.test_document_comment.text, comment_text)
 
     def test_comment_list_view_with_no_permission(self):
         self._create_test_comment()
 
         response = self._request_test_comment_list_view()
         self.assertNotContains(
-            response=response, text=self.test_document_comment.comment,
+            response=response, text=self.test_document_comment.text,
             status_code=404
         )
 
@@ -192,7 +192,7 @@ class DocumentCommentViewTestCase(
 
         response = self._request_test_comment_list_view()
         self.assertContains(
-            response=response, text=self.test_document_comment.comment,
+            response=response, text=self.test_document_comment.text,
             status_code=200
         )